+++ src/main/java/kimp/chat/service/serviceImpl/ChatServiceImpl.java
@@ -1,6 +1,6 @@
 package kimp.chat.service.serviceImpl;
 
-import kimp.chat.dao.daoImpl.ChatDao;
+import kimp.chat.dao.ChatDao;
 import kimp.chat.dto.response.ChatLogResponseDto;
 import kimp.chat.entity.Chat;
 import kimp.chat.service.ChatService;
